fix(demo): prevent events to propagate from UI elements to itowns view
- made FeaturePickerService's blockEvents method a global utils so PointCloudScene's UI can use it
- check skymanager's existence and enabled property before define
